+++ RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+#### 0.9.9
+* Code documentation
+
+#### 0.9.4
+* Use PassGenerics on client CE-API
+
+#### 0.9.3
+* More consistent API
+
+#### 0.9.2
+* Test new API
+
+#### 0.9.0
+* Project renamed to Elmish.Bridge
+
 #### 0.7.5
 * Support longer messages